Default value for MXMC counters in the output TTree
Previously didn't consider the case in which the number of integration dimensions is zero (although the chosen integrator was Markov Chain). Also, if there was an error during the MEM score evaluation, the MXMC counter can be garbage. Also encountered a type conversion mismatch which wasn't a bug in itself (b/c long long covers unsigned type) but still... Forgot to comment some code in the previous commit.

+++ interface/MEMIntegratorMarkovChain.h
@@ -167,7 +167,7 @@ namespace tthMEM
     unsigned long nofMoves_accepted_,      nofMoves_rejected_,
                   nofMoves_acceptedTotal_, nofMoves_rejectedTotal_,
                   nofIntegrationCalls_;
-    unsigned nofTries_;
+    long long nofTries_;
     ///< number of calls made until positive probability has been found (summed over chains)
 
     TRandom3 prng_;
